Contrato API Contrato AsignarAcopio

+++ CoffeeConnect.Concrete/PrecioDiaRendimientoService.cs
@@ -56,11 +56,11 @@ namespace CoffeeConnect.Service
         {
             CalculoPrecioDiaRendimientoDTO calculoPrecioDiaRendimientoDTO = new CalculoPrecioDiaRendimientoDTO();
                         
-            double precioPromedio = _IContratoRepository.CalcularPrecioDiaContrato(request.EmpresaId);
+            decimal precioPromedio = _IContratoRepository.CalcularPrecioDiaContrato(request.EmpresaId);
 
             calculoPrecioDiaRendimientoDTO.PrecioPromedioContrato = precioPromedio;
 
-            calculoPrecioDiaRendimientoDTO.TipoCambio = 4.1;
+            calculoPrecioDiaRendimientoDTO.TipoCambio = 4.1M;
 
             List<CalculoPrecioDiaRendimientoBE> precios = new List<CalculoPrecioDiaRendimientoBE>();
 
@@ -68,19 +68,19 @@ namespace CoffeeConnect.Service
 
             CalculoPrecioDiaRendimientoBE precio1 = new CalculoPrecioDiaRendimientoBE();                 
             precio1.RendimientoInicio = 64;
-            precio1.RendimientoFin = 65.99;
+            precio1.RendimientoFin = 65.99M;
             precio1.KGPergamino = valorConstante/(precio1.RendimientoInicio/100);
             precio1.PrecioDia = precioPromedio / precio1.KGPergamino;
 
             CalculoPrecioDiaRendimientoBE precio2 = new CalculoPrecioDiaRendimientoBE();
             precio2.RendimientoInicio = 66;
-            precio2.RendimientoFin = 67.99;
+            precio2.RendimientoFin = 67.99M;
             precio2.KGPergamino = valorConstante / (precio2.RendimientoInicio / 100);
             precio2.PrecioDia = precioPromedio / precio2.KGPergamino;
 
             CalculoPrecioDiaRendimientoBE precio3 = new CalculoPrecioDiaRendimientoBE();           
             precio3.RendimientoInicio = 68;
-            precio3.RendimientoFin = 69.99;
+            precio3.RendimientoFin = 69.99M;
             precio3.KGPergamino = valorConstante / (precio3.RendimientoInicio / 100);
             precio3.PrecioDia = precioPromedio / precio3.KGPergamino;
 
@@ -88,20 +88,20 @@ namespace CoffeeConnect.Service
 
             CalculoPrecioDiaRendimientoBE precio4 = new CalculoPrecioDiaRendimientoBE();         
             precio4.RendimientoInicio = 70;
-            precio4.RendimientoFin = 71.99;
+            precio4.RendimientoFin = 71.99M;
             precio4.KGPergamino = valorConstante / (precio4.RendimientoInicio / 100);
             precio4.PrecioDia = precioPromedio / precio4.KGPergamino;
 
 
             CalculoPrecioDiaRendimientoBE precio5 = new CalculoPrecioDiaRendimientoBE();
             precio5.RendimientoInicio = 72;
-            precio5.RendimientoFin = 73.99;
+            precio5.RendimientoFin = 73.99M;
             precio5.KGPergamino = valorConstante / (precio5.RendimientoInicio / 100);
             precio5.PrecioDia = precioPromedio / precio5.KGPergamino;
 
             CalculoPrecioDiaRendimientoBE precio6 = new CalculoPrecioDiaRendimientoBE();        
             precio6.RendimientoInicio = 74;
-            precio6.RendimientoFin = 75.99;
+            precio6.RendimientoFin = 75.99M;
             precio6.KGPergamino = valorConstante / (precio6.RendimientoInicio / 100);
             precio6.PrecioDia = precioPromedio / precio6.KGPergamino;
 
@@ -109,13 +109,13 @@ namespace CoffeeConnect.Service
 
             CalculoPrecioDiaRendimientoBE precio7 = new CalculoPrecioDiaRendimientoBE();         
             precio7.RendimientoInicio = 76;
-            precio7.RendimientoFin = 77.99;
+            precio7.RendimientoFin = 77.99M;
             precio7.KGPergamino = valorConstante / (precio7.RendimientoInicio / 100);
             precio7.PrecioDia = precioPromedio / precio7.KGPergamino;
 
             CalculoPrecioDiaRendimientoBE precio8 = new CalculoPrecioDiaRendimientoBE();     
             precio8.RendimientoInicio = 78;
-            precio8.RendimientoFin = 79.99;
+            precio8.RendimientoFin = 79.99M;
             precio8.KGPergamino = valorConstante / (precio8.RendimientoInicio / 100);
             precio8.PrecioDia = precioPromedio / precio8.KGPergamino;
 
@@ -153,36 +153,36 @@ namespace CoffeeConnect.Service
 
             PorcentajeRendimientoBE rendimiento1 = new PorcentajeRendimientoBE();
             rendimiento1.RendimientoInicio = 64;
-            rendimiento1.RendimientoFin = 65.99;
+            rendimiento1.RendimientoFin = 65.99M;
             rendimiento1.KGPergamino = valorConstante / (rendimiento1.RendimientoInicio / 100);
             
 
             PorcentajeRendimientoBE rendimiento2 = new PorcentajeRendimientoBE();
             rendimiento2.RendimientoInicio = 66;
-            rendimiento2.RendimientoFin = 67.99;
+            rendimiento2.RendimientoFin = 67.99M;
             rendimiento2.KGPergamino = valorConstante / (rendimiento2.RendimientoInicio / 100);
            
 
             PorcentajeRendimientoBE rendimiento3 = new PorcentajeRendimientoBE();
             rendimiento3.RendimientoInicio = 68;
-            rendimiento3.RendimientoFin = 69.99;
+            rendimiento3.RendimientoFin = 69.99M;
             rendimiento3.KGPergamino = valorConstante / (rendimiento3.RendimientoInicio / 100);
 
             PorcentajeRendimientoBE rendimiento4 = new PorcentajeRendimientoBE();
             rendimiento4.RendimientoInicio = 70;
-            rendimiento4.RendimientoFin = 71.99;
+            rendimiento4.RendimientoFin = 71.99M;
             rendimiento4.KGPergamino = valorConstante / (rendimiento4.RendimientoInicio / 100);
            
 
             PorcentajeRendimientoBE rendimiento5 = new PorcentajeRendimientoBE();
             rendimiento5.RendimientoInicio = 72;
-            rendimiento5.RendimientoFin = 73.99;
+            rendimiento5.RendimientoFin = 73.99M;
             rendimiento5.KGPergamino = valorConstante / (rendimiento5.RendimientoInicio / 100);
             
 
             PorcentajeRendimientoBE rendimiento6 = new PorcentajeRendimientoBE();
             rendimiento6.RendimientoInicio = 74;
-            rendimiento6.RendimientoFin = 75.99;
+            rendimiento6.RendimientoFin = 75.99M;
             rendimiento6.KGPergamino = valorConstante / (rendimiento6.RendimientoInicio / 100);
             
 
@@ -190,13 +190,13 @@ namespace CoffeeConnect.Service
 
             PorcentajeRendimientoBE rendimiento7 = new PorcentajeRendimientoBE();
             rendimiento7.RendimientoInicio = 76;
-            rendimiento7.RendimientoFin = 77.99;
+            rendimiento7.RendimientoFin = 77.99M;
             rendimiento7.KGPergamino = valorConstante / (rendimiento7.RendimientoInicio / 100);
             
 
             PorcentajeRendimientoBE rendimiento8 = new PorcentajeRendimientoBE();
             rendimiento8.RendimientoInicio = 78;
-            rendimiento8.RendimientoFin = 79.99;
+            rendimiento8.RendimientoFin = 79.99M;
             rendimiento8.KGPergamino = valorConstante / (rendimiento8.RendimientoInicio / 100);
            
             PorcentajeRendimientoBE rendimiento9 = new PorcentajeRendimientoBE();
