+++ CoffeeConnect.ViewModels/PrecioDiaRendimiento/CalculoPrecioDiaRendimientoBE.cs
@@ -11,10 +11,10 @@ namespace CoffeeConnect.DTO
 
         }
 
-        public double RendimientoInicio { get; set; }
-        public double RendimientoFin { get; set; }
-        public double KGPergamino { get; set; }
-        public double PrecioDia { get; set; }
+        public decimal RendimientoInicio { get; set; }
+        public decimal RendimientoFin { get; set; }
+        public decimal KGPergamino { get; set; }
+        public decimal PrecioDia { get; set; }
      
         
     }
