+++ CoffeeConnect.ViewModels/PrecioDiaRendimiento/PorcentajeRendimientoBE.cs
@@ -11,9 +11,9 @@ namespace CoffeeConnect.DTO
 
         }
 
-        public double RendimientoInicio { get; set; }
-        public double RendimientoFin { get; set; }
-        public double KGPergamino { get; set; }
+        public decimal RendimientoInicio { get; set; }
+        public decimal RendimientoFin { get; set; }
+        public decimal KGPergamino { get; set; }
